feat: NetworkError 구현
