+++ BoxOffice/Error/NetworkError.swift
@@ -1,0 +1,12 @@
+//
+//  NetworkError.swift
+//  BoxOffice
+//
+//  Created by vetto, brody on 23/03/21.
+//
+
+enum NetworkError: Error {
+    case clientError
+    case serverError
+    case unknownError
+}
